TriggerBase: Unit tests

+++ include/examples/battery_health_trigger.hpp
@@ -10,7 +10,7 @@ namespace ros2bag_triggered::examples
 class BatteryHealthTrigger : public TriggerBase<sensor_msgs::msg::BatteryState>
 {
 public:
-    explicit BatteryHealthTrigger(uint64_t persistance_duration, const rclcpp::Clock::SharedPtr clock, const std::shared_ptr<rclcpp::Logger> logger, bool use_msg_stamp)
+    explicit BatteryHealthTrigger(double persistance_duration, const rclcpp::Clock::SharedPtr clock, const std::shared_ptr<rclcpp::Logger> logger, bool use_msg_stamp)
     : TriggerBase(persistance_duration, clock, logger, use_msg_stamp) {}
 
     BatteryHealthTrigger() = delete;
